--- packages/http/src/base/index.ts
@@ -1,2 +1,0 @@
-export { HttpBase, RequestInitBase, RequestInitGet } from './base'
-export { HttpJson } from '../content-type/json'
