+++ packages/http/src/content-type/http_error.ts
@@ -1,0 +1,5 @@
+import { Response } from 'node-fetch'
+
+export interface HttpError extends Response {
+  message: string
+}
