+++ packages/http/src/index.ts
@@ -1,9 +1,13 @@
-export { RequestInitBase, RequestInitGet } from './base'
+export { RequestInitBase, RequestInitGet } from './http'
 export { RequestInit, Response } from 'node-fetch'
 
-export { Http } from './cloud'
-export { HttpSeed } from './seed'
-export { HttpPure } from './pure'
+// mode provider
+export { httpModeProvider } from './provider'
 
-// typed
-export { HttpJson } from './content-type/json'
+// http injectables
+export { Http } from './modes/cloud'
+export { HttpPure } from './modes/pure'
+export { HttpSeed } from './modes/seed'
+
+// content-type clients (convinience)
+export { HttpJson } from './content-type/http_json'
