+++ packages/http/src/modes/cloud.ts
@@ -1,7 +1,7 @@
 import { Injectable } from 'injection-js'
 import fetch, { RequestInit, Response } from 'node-fetch'
 import 'reflect-metadata'
-import { HttpBase } from './base'
+import { HttpBase } from '../base'
 
 /**
  * Http (default) cloud mode implementation.
