+++ packages/http/test/cloud.ts
@@ -6,6 +6,12 @@ import { skip, slow, suite, test, timeout } from 'mocha-typescript'
 import * as querystring from 'querystring'
 import { Http } from '../src'
 
+interface ExpectedResponse {
+  args: {
+    param1: string
+    param2: string
+  }
+}
 @suite('Http::Cloud::Fetch')
 export class Test extends BasicTest {
   injector = ReflectiveInjector.resolveAndCreate([Http])
