--- packages/http/test/index.ts
@@ -1,6 +1,0 @@
-interface ExpectedResponse {
-  args: {
-    param1: string
-    param2: string
-  }
-}
